Misc: Move BeatmapHelpers to Types project
