+++ plugins/BeatmapPpReplacements/BeatmapPpReplacements.csproj
@@ -47,7 +47,6 @@
     <Reference Include="System.Xml" />
   </ItemGroup>
   <ItemGroup>
-    <Compile Include="BeatmapHelpers.cs" />
     <Compile Include="PpReplacements.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
   </ItemGroup>
